Revert "Merge branch 'master' into resolution-rescaling-4"
This reverts commit f4b612e98e1df4017c33741fc8d77d1290286eb8.

+++ src/core/file_sys/kernel_executable.cpp
@@ -147,7 +147,7 @@ std::vector<u32> KIP::GetKernelCapabilities() const {
 }
 
 s32 KIP::GetMainThreadPriority() const {
-    return static_cast<s32>(header.main_thread_priority);
+    return header.main_thread_priority;
 }
 
 u32 KIP::GetMainThreadStackSize() const {
